#3.1 Ride endpoints
Created endpoint: creating a ride

+++ src/main/java/rs/ac/uns/ftn/informatika/jpa/model/Admin.java
@@ -7,11 +7,11 @@ public class Admin extends User{
     public Admin() {
     }
 
-    public Admin(int id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active) {
+    public Admin(Long id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active) {
         super(id, firstName, lastName, picture, phoneNumber, email, address, password, blocked, active);
     }
 
-    public Admin(int id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active, String username) {
+    public Admin(Long id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active, String username) {
         super(id, firstName, lastName, picture, phoneNumber, email, address, password, blocked, active);
         this.username = username;
     }
