+++ src/main/java/rs/ac/uns/ftn/informatika/jpa/model/Driver.java
@@ -8,15 +8,15 @@ public class Driver extends User{
     ArrayList<Ride> rides;
     Vehicle vehicle;
 
-    public Driver(int id, String email) {
+    public Driver(Long id, String email) {
         super(id, email);
     }
 
-    public Driver(int id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active) {
+    public Driver(Long id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active) {
         super(id, firstName, lastName, picture, phoneNumber, email, address, password, blocked, active);
     }
 
-    public Driver(int id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active, ArrayList<Document> documents, ArrayList<Ride> rides, Vehicle vehicle) {
+    public Driver(Long id, String firstName, String lastName, String picture, String phoneNumber, String email, String address, String password, boolean blocked, boolean active, ArrayList<Document> documents, ArrayList<Ride> rides, Vehicle vehicle) {
         super(id, firstName, lastName, picture, phoneNumber, email, address, password, blocked, active);
         this.documents = documents;
         this.rides = rides;
